+++ Cargo.toml
@@ -7,5 +7,6 @@ edition = "2021"
 
 [dependencies]
 fmt = "0.1.0"
+rand = "0.8.5"
 serde = { version = "1.0.195", features = ["derive"] }
 toml = "0.8.8"
