+++ src/main.rs
@@ -2,7 +2,7 @@ use std::path::Path;
 
 mod gamemaker;
 mod gamemakerv2;
-mod riddler;
+// mod riddler;
 
 fn main() {
     gamemakerv2::main_loop();
